Add First Checked System.out.println For Failed Parsing
Added a print statement in JavaCompiler.java in the compiler method.
The statement is run when the 'todo' variable is empty. This looks to
be empty when the previous stages have failed (in this case, only parsing).
Also added a print statement to print the compile policy - for debugging.
Compile policy should be BY_TODO to run the parsing check and print statement.
Issue #4